requirements.txt
# Core LLM & RAG
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Vector stores
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Embeddings & NLP
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.2.0
tiktoken>=0.7.0

# Knowledge graph
networkx>=3.3
spacy>=3.7.0
# python -m spacy download en_core_web_sm

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Data handling
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0

# Evaluation
ragas>=0.1.14
bert-score>=0.3.13
rouge-score>=0.1.2
nltk>=3.8.1

# App / UI
streamlit>=1.35.0
gradio>=4.36.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
httpx>=0.27.0

# Local LLM (optional)
ollama>=0.2.0
